// src/features/compiler/compiler-core-assets.generated.ts
// Generated by scripts/prepare-compiler-assets.mjs. Do not edit.

export type CoreAssetId =
  | "busytex-worker"
  | "busytex-pipeline"
  | "busytex-js"
  | "busytex-wasm"
  | "texlive-basic-js"
  | "texlive-basic-data";

export interface GeneratedCoreAsset {
  id: CoreAssetId;
  filename: string;
  byteLength: number;
  sha256: string;
  mimeType: string;
}

export const GENERATED_CORE_ASSETS: readonly GeneratedCoreAsset[] =
  [
  {
    "id": "busytex-worker",
    "mimeType": "text/javascript",
    "filename": "busytex_worker.js",
    "byteLength": 3006,
    "sha256": "c27e7f7e5772984a2348d4381ef51122b75fc820d15ccb3b2c69a9fa5d256a0c"
  },
  {
    "id": "busytex-pipeline",
    "mimeType": "text/javascript",
    "filename": "busytex_pipeline.js",
    "byteLength": 33701,
    "sha256": "5fea68b97caa45cfd17899903811442fc0310b50888ef3787d3ac3457ed5fced"
  },
  {
    "id": "busytex-js",
    "mimeType": "text/javascript",
    "filename": "busytex.js",
    "byteLength": 271047,
    "sha256": "316a48050159ce60d57aa58d66b5cf3ca7a8f13525179f1d2dd56ed095a49819"
  },
  {
    "id": "busytex-wasm",
    "mimeType": "application/wasm",
    "filename": "busytex.wasm",
    "byteLength": 32501975,
    "sha256": "58d60c73c2aa4fc09ae86d671d13e083c5cd7c34ab8878f42abba29a1cf0348f"
  },
  {
    "id": "texlive-basic-js",
    "mimeType": "text/javascript",
    "filename": "texlive-basic.js",
    "byteLength": 2049150,
    "sha256": "9b66d750257045f5c6a8b6c35266608c4dc156e8ea50a7b575327c0f43a0943a"
  },
  {
    "id": "texlive-basic-data",
    "mimeType": "application/octet-stream",
    "filename": "texlive-basic.data",
    "byteLength": 90786746,
    "sha256": "a5be427683774e8de6d6f3ea6b8255758eb973a4dac5df86a3f4cf5a837ccf81"
  }
] as const;